frontend/requirement.txt
numpy
pandas
scipy
shapely
geopandas
pymongo
python-dotenv
requests
